refactors code for device ID auth
