+++ src/types/device.ts
@@ -1,0 +1,5 @@
+/** Represents a device id pair */
+export type DeviceIdPair = {
+    realDeviceId: string,
+    storedDeviceId: string
+}
