+++ src/utils/deviceId/deviceIdUtils.ts
@@ -3,7 +3,10 @@
  */
 
 import { DjangoAuthUser } from '../../types/auth';
+import { DeviceIdPair } from '../../types/device';
 import { DeviceIdUnavailableError } from './exceptions';
+import { getMobileDeviceId } from './mobileDeviceIdUtils';
+import { getWebDeviceId } from './webDeviceIdUtils';
 
 /**
  * Error thrown when a device ID cannot be obtained on mobile.
@@ -108,52 +111,14 @@ function normalizeDeviceIdStrict(id: string): string {
  * Get the device ID - either from mobile app (strict) or generate/retrieve for web
  * @returns Promise<string> The device ID
  */
-export async function getDeviceId(): Promise<{ realDeviceId: string; storedDeviceId: string }> {
-  // Mobile: must return native device ID or throw. No UUID fallback.
-
-  //localStorage.setItem('device_id', 'test');
+export async function getDeviceId(): Promise<DeviceIdPair> {
 
   if (isMobileApp()) {
-    console.log('Mobile detected...');
-
-    const storedDeviceId = localStorage.getItem('device_id') || '';
-    const realDeviceId = window.deviceId || '';
-
-    console.log('Stored device ID: ', storedDeviceId);
-    console.log('Real device ID: ', realDeviceId);
-
-    if (!storedDeviceId && !realDeviceId) {
-      throw new DeviceIdUnavailableError('Unable to obtain mobile device ID');
-    }
-
-    return { realDeviceId, storedDeviceId };
+    return getMobileDeviceId();
   }
 
-  console.log('Web detected; using deviceId from localStorage or creating a new one...');
+  return getWebDeviceId();
 
-  // Web browser: use localStorage
-  let webDeviceId = localStorage.getItem('device_id');
-
-  if (!webDeviceId) {
-    // Generate a new device ID
-    if (
-      typeof crypto !== 'undefined' &&
-      (crypto as Crypto & { randomUUID?: () => string }).randomUUID
-    ) {
-      webDeviceId = (crypto as Crypto & { randomUUID: () => string }).randomUUID();
-    } else {
-      // Fallback for older browsers
-      webDeviceId = `web_${Date.now()}_${Math.random().toString(36).substring(2, 15)}`;
-    }
-
-    localStorage.setItem('device_id', webDeviceId);
-    console.log('New device ID for web: ', webDeviceId);
-  }
-
-  console.log('deviceId exists in local storage: ', webDeviceId);
-
-  // For web, both real and stored device IDs are the same
-  return { realDeviceId: webDeviceId, storedDeviceId: webDeviceId };
 }
 
 /**
